tests: cleanup tests to avoid unneeded imports
