+++ psycop/common/sequence_models/tests/conftest.py
@@ -1,0 +1,39 @@
+from datetime import datetime
+
+import pytest
+
+from psycop.common.data_structures import Patient, TemporalEvent
+
+
+@pytest.fixture()
+def patients() -> list[Patient]:
+    """
+    Returns a list of patient objects
+    """
+
+    e1 = TemporalEvent(
+        timestamp=datetime(2021, 1, 1),
+        value="d1",
+        source_type="diagnosis",
+        source_subtype="",
+    )
+    e2 = TemporalEvent(
+        timestamp=datetime(2021, 1, 3),
+        value="d2",
+        source_type="diagnosis",
+        source_subtype="",
+    )
+
+    patient1 = Patient(
+        patient_id=1,
+        date_of_birth=datetime(1990, 1, 1),
+    )
+    patient1.add_events([e1, e2])
+
+    patient2 = Patient(
+        patient_id=2,
+        date_of_birth=datetime(1993, 3, 1),
+    )
+    patient2.add_events([e1, e2, e2, e1])
+
+    return [patient1, patient2]
